requirements.txt
httpx==0.26.0
pyyaml==6.0.1
pandas==2.1.4
pyarrow==14.0.2
duckdb==0.9.2